test(storage): fix copy_configuration script (#3635)

+++ AmplifyPlugins/Storage/Tests/StorageHostApp/copy_configuration.sh
@@ -25,6 +25,13 @@ else
     touch "$DESTINATION_DIR/amplifyconfiguration.json"
 fi
 
+if [ -f "$SOURCE_DIR/AWSS3StoragePluginTests-amplifyconfiguration.json" ]; then
+    cp "$SOURCE_DIR/AWSS3StoragePluginTests-amplifyconfiguration.json" "$DESTINATION_DIR/amplifyconfiguration.json"
+    exit 0
+fi
+    touch "$DESTINATION_DIR/amplifyconfiguration.json"
+fi
+
 if [ -f "$SOURCE_DIR/AWSS3StoragePluginTests-amplify_outputs.json" ]; then
     cp "$SOURCE_DIR/AWSS3StoragePluginTests-amplify_outputs.json" "$DESTINATION_DIR/amplify_outputs.json"
     exit 0
